+++ packages/browser/src/queues.ts
@@ -42,6 +42,9 @@ export class QueuesStats {
     if (!this._opt.performanceStats) {
       return;
     }
+    if (!this._opt.queueStats) {
+      return;
+    }
 
     let ms = q._duration();
     if (ms === 0) {